requirements.txt
django>=4.*
pytest-django==4.5.2
pytest-mock==3.8.2
tox==3.25.1
